requirements.txt
Django~=2.1.0
djangorestframework~=3.8.2
django-easy-pdf
WeasyPrint>=0.34
